fix: Using validators as opposed to accept prop
The accept prop doesn't seem to properly deal with weird file types like
the Satisfactory save file since it wants proper mime types. Instead, we
are using a validator that will simply check the file extension. Since
we're dealing with weird binary files, this is bound to be sort of
weird.